+++ firebase/functions-admin/src/index.ts
@@ -1,0 +1,6 @@
+import { initializeApp } from 'firebase-admin/app';
+
+initializeApp();
+
+export { onUserCreate } from './on-user-create';
+export { claimAdmin } from './claim-admin';